+++ lib/llm/providers/bedrock/response_adapter/completion.rb
@@ -57,6 +57,18 @@ module LLM::Bedrock::ResponseAdapter
     end
 
     ##
+    # (see LLM::Contract::Completion#input_audio_tokens)
+    def input_audio_tokens
+      super
+    end
+
+    ##
+    # (see LLM::Contract::Completion#output_audio_tokens)
+    def output_audio_tokens
+      super
+    end
+
+    ##
     # (see LLM::Contract::Completion#cache_read_tokens)
     def cache_read_tokens
       0