+++ lib/llm/providers/google/response_adapter/completion.rb
@@ -34,6 +34,18 @@ module LLM::Google::ResponseAdapter
     end
 
     ##
+    # (see LLM::Contract::Completion#input_audio_tokens)
+    def input_audio_tokens
+      super
+    end
+
+    ##
+    # (see LLM::Contract::Completion#output_audio_tokens)
+    def output_audio_tokens
+      super
+    end
+
+    ##
     # (see LLM::Contract::Completion#cache_read_tokens)
     def cache_read_tokens
       0